+++ Paylocity.CodingChallenge.Business/Code/Constants.cs
@@ -8,9 +8,9 @@ namespace Paylocity.CodingChallenge.Business.Code
     [ExcludeFromCodeCoverage]
     public class Constants
     {
-        public const double EMPLOYEE_DEDUCTION_PER_YEAR = 1000;
-        public const double DEPENDENT_DEDUCTION_PER_YEAR = 500;
-        public const double TEN_PERCENT_DISCOUNT_RATE = 0.10;
-        public const double ZERO_PERCENT_DISCOUNT_RATE = 0.0;
+        public const decimal EMPLOYEE_DEDUCTION_PER_YEAR = 1000;
+        public const decimal DEPENDENT_DEDUCTION_PER_YEAR = 500;
+        public const decimal TEN_PERCENT_DISCOUNT_RATE = 0.10M;
+        public const decimal ZERO_PERCENT_DISCOUNT_RATE = 0.0M;
     }
 }
